+++ src/components/index.ts
@@ -8,6 +8,6 @@ export { default as Rating } from './Rating/Rating';
 export { default as Modal } from './Modal/Modal';
 export { default as ImageInput } from './WriteEditModal/ImageInput';
 export { default as KeyboardInfoBox } from './WriteEditModal/KeyboardInfoBox';
-export { default as WriteEditModal } from './WriteEditModal/WriteEditModal';
 
+export { default as Button } from './Button/Button';
 // export { default as ItemOverview } from './ItemOverview/ItemOverview';
